found & fixed some holes in files that weren't open
