+++ src/renderer/components/content-types/files/AudioContent.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { FileDoc } from './FileContent'
+import { FileDoc } from '.'
 
 import { ContentProps } from '../../Content'
 import ContentTypes from '../../../ContentTypes'
