+++ src/renderer/components/content-types/files/FileInList.css
@@ -1,0 +1,11 @@
+.FileListItem {
+  display: flex;
+  align-items: center;
+  width: 100%;
+}
+
+.FileListItem__title {
+  margin-left: 10px;
+  -webkit-user-drag: none;
+  user-select: none;
+}
