+++ src/renderer/components/content-types/files/PdfContent.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useCallback, useMemo } from 'react'
 
 import { Document, Page } from 'react-pdf/dist/entry.webpack'
-import { FileDoc } from './FileContent'
+import { FileDoc } from '.'
 
 import ContentTypes from '../../../ContentTypes'
 import { ContentProps } from '../../Content'
